% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_heatMap}
\alias{create_heatMap}
\title{create heatmap of user's home and office}
\usage{
create_heatMap(loc, shop)
}
\arguments{
\item{loc}{the dataset which is produced by detectOffice}

\item{shop}{the shop address dataset read by readShop function}
}
\description{
create heatmap of user's home and office
}

