% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_pathPlot}
\alias{create_pathPlot}
\title{visualize an individual's trajectory}
\usage{
create_pathPlot(id1, loc)
}
\arguments{
\item{loc}{the origin dataset which is read by readLoc function}

\item{id}{the user's id}
}
\description{
visualize an individual's trajectory
}

