% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_pathPlots}
\alias{create_pathPlots}
\title{visualize a group of users' trajectory}
\usage{
create_pathPlots(ids, loc, address, suffix = "")
}
\arguments{
\item{loc}{the origin dataset which is read by readLoc function}

\item{id}{the users' ids}
}
\description{
visualize a group of users' trajectory
}
\details{
avoiding two much data in html file, I recommend no more than 1000 users's trajectory to be visualized
}

