% Generated by roxygen2 (4.0.2): do not edit by hand
\name{detectOffice}
\alias{detectOffice}
\title{detect office address and home address}
\usage{
detectOffice(data)
}
\arguments{
\item{data}{the dataset produced by `findTwoPlaces` function}
}
\description{
distinguish between two places, detect which is office and which is home
}

