% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findTwoPlaces}
\alias{findTwoPlaces}
\title{find two places where the user appeared most frequently}
\usage{
findTwoPlaces(data, dis_thres)
}
\arguments{
\item{data}{the dataset processed by function `handleLoc`}

\item{dis_thres}{minimum distance between the two places where the user appeared most frequently}
}
\description{
find two places where the user appeared most frequently
}

