% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getAddress}
\alias{getAddress}
\title{get address information via longitude and latitude}
\usage{
getAddress(lon, lat)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}
}
\description{
get address information via longitude and latitude
}
\details{
the output is a list which contains four elements:detailed address, province, district, street
}

