% Generated by roxygen2 (4.0.2): do not edit by hand
\name{handleLoc}
\alias{handleLoc}
\title{handle the location information}
\usage{
handleLoc(data)
}
\arguments{
\item{data}{the dataset read by read.loc}
}
\description{
reduce the redundant location information and calculate the time a user stayed in one place
}

