% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{loc_cache_show}
\alias{loc_cache_show}
\title{Show the session cache}
\usage{
loc_cache_show()
}
\value{
a data frame of all cached results, or an empty data frame if the
cache is empty.
}
\description{
Show the session cache
}
